% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{cents2frac}
\alias{cents2frac}
\title{Cents to Fraction}
\usage{
cents2frac(cents)
}
\arguments{
\item{cents}{a numeric vector of cents values}
}
\value{
a vector of the corresponding vulgar fractions
}
\description{
Converts a vector of cents to a vector of the corresponding
ratios, expressed as a character vector of vulgar fractions
}
\examples{
  print(cents2frac(seq(0, 1200, by = 100)))

}
