% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonaR.R
\name{cents2ratio}
\alias{cents2ratio}
\title{Cents to Ratio}
\usage{
cents2ratio(cents)
}
\arguments{
\item{cents}{a numeric vector of cents values}
}
\value{
a numeric vector of the corresponding ratios
}
\description{
Converts a vector of cents to a vector of the corresponding
ratios
}
\examples{
  print(cents2ratio(seq(0, 1200, by = 100)))

}
