% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonaR.R
\name{et_scale_table}
\alias{et_scale_table}
\title{Create Equal-Tempered Scale Table}
\usage{
et_scale_table(period = 2, divisions = 12, tonic_note_number = 60)
}
\arguments{
\item{period}{The period - default is 2, for an octave}

\item{divisions}{Number of degrees in the scale - default is 12}

\item{tonic_note_number}{MIDI note number of the tonic for the scale -
\itemize{
\item default is middle C = 60
}}
}
\value{
a \code{data.table} with seven columns:
\itemize{
\item \code{ratio}: the ratio that defines the note, as a number between 1 and
\code{period}
\item \code{ratio_frac}: the ratio as a vulgar fraction (character). The ratios
for this type of scale are usually irrational, so this is an approximation,
computed by \code{fractional::fractional}.
\item \code{ratio_cents}: the ratio in cents (hundredths of a semitone)
\item \code{frequency}: frequency of the note given the \code{tonic_note_number}
parameter
\item \code{bent_midi}: the MIDI note number as an integer plus a fraction. For
example, middle C is MIDI note number 60 and middle C sharp is 61. The
quarter-tone half-way between C and C sharp would have a \code{bent_midi} value
of 60.5. The name \code{bent_midi} comes from the fact that a MIDI sequencer
can convert the value to a regular integer MIDI note number message and
a pitch bend message.
\item \code{interval_cents}: interval between this note and the previous note
\item \code{degree}: scale degree from zero to (number of notes) - 1
}
}
\description{
Creates a scale table for equal divisions of a specified
period.
}
\examples{

print(vanilla <- et_scale_table()) # default is 12EDO, of course

# 19-EDO
print(edo19 <- et_scale_table(2.0, 19))

# 31-EDO
print(edo31 <- et_scale_table(2.0, 31))

# equal-tempered Bohlen-Pierce
print(bohlen_pierce_et <- et_scale_table(3.0, 13))

# Carlos Alpha
print(carlos_alpha <- et_scale_table(1.5, 9))
}
