% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{freq2MIDI}
\alias{freq2MIDI}
\title{Frequency to MIDI note number}
\usage{
freq2MIDI(freq)
}
\arguments{
\item{freq}{a numeric vector of frequency values}
}
\value{
a numeric vector of the corresponding MIDI note numbers
}
\description{
Converts a vector of frequencies to a vector of the corresponding
MIDI note numbers. Note that these MIDI note numbers are \emph{\strong{floating point}}
values - for example, MIDI note number 60.5 is the quarter-tone between
middle C (60) and middle C# (61).
}
\examples{

  # chromatic scale
  freq <- 2 ^ (seq(0, 12) / 12) * 220
  print(freq2MIDI(freq))

}
