% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonaR.R
\name{MIDI2freq}
\alias{MIDI2freq}
\title{MIDI note number to frequency}
\usage{
MIDI2freq(MIDI)
}
\arguments{
\item{MIDI}{a numeric vector of MIDI note numbers}
}
\value{
a numeric vector of the corresponding frequencies
}
\description{
Converts a vector of MIDI note numbers to a vector of the
corresponding frequencies. Note that these MIDI note numbers are
\emph{\strong{floating point}}values - for example, MIDI note number 60.5 is the
quarter-tone between middle C (60) and middle C# (61).
}
\examples{

  # quarter-tone scale
  MIDI <- seq(60, 72, by = 0.5)
  print(MIDI2freq(MIDI))

}
