% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{period_reduce}
\alias{period_reduce}
\title{Period Reduce}
\usage{
period_reduce(scale, period = 2)
}
\arguments{
\item{scale}{a numeric vector of ratios representing a scale}

\item{period}{the period of the scale - default is 2.0, the "octave"}
}
\value{
the scale reduced so that all ratios are between 1 and the period
}
\description{
Reduce a vector of ratios to a given period
}
\examples{
  test_scale <- cents2ratio(seq(2400, 3500, by = 100))
  print(scale <- period_reduce(test_scale, 2))

}
