% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ratio2cents}
\alias{ratio2cents}
\title{Ratio to Cents}
\usage{
ratio2cents(ratio)
}
\arguments{
\item{ratio}{a numeric vector of ratios}
}
\value{
a numeric vector of the corresponding cents values
}
\description{
Converts a vector of ratios to a vector of the corresponding
cents values
}
\examples{
  i <- seq(0, 12)
  print(ratio2cents(2.0 ^ (i / 12.0)))

}
