% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonaR.R
\name{ratio2factors}
\alias{ratio2factors}
\title{Ratio to Factors}
\usage{
ratio2factors(ratio)
}
\arguments{
\item{ratio}{a numeric vector of ratios}
}
\value{
a list of vectors, with each vector the integer factors that
determined the corresponding ratio
}
\description{
Converts a vector of ratios to a vector of the corresponding
integer factors.
}
\examples{
(super <- sclfile_scale_table(system.file(
  "test_scl_files/carlos_super.scl",
  package = "consonaR"
))$scale_table$ratio)
(super_factors <- ratio2factors(super))

(harm <- sclfile_scale_table(system.file(
  "test_scl_files/carlos_harm.scl",
  package = "consonaR"
))$scale_table$ratio)
(harm_factors <- ratio2factors(harm))

# we know the factors that yield the 1-3-5-7 Hexany
# can we recover them?
(hexany <- prodset_scale_table(list(
  c(1, 3),
  c(1, 5),
  c(1, 7),
  c(3, 5),
  c(3, 7),
  c(5, 7)
))$ratio)
(hexany_factors <- ratio2factors(hexany))

}
