% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{sclfile_scale_table}
\alias{sclfile_scale_table}
\title{Create Scale Table From a Scala \code{.scl} File}
\usage{
sclfile_scale_table(sclfile_path, root_freq = 440/(2^(9/12)))
}
\arguments{
\item{sclfile_path}{The path to a valid Scala \code{.scl} file}

\item{root_freq}{root frequency of the scale - default is middle C:
440 / (2 ^ (9 / 12))}
}
\value{
a list with two or three items:
\itemize{
\item \code{status} (character): "Oll Korrect" if the results are valid,
otherwise an error message
\item \code{file_contents} (character vector): the contents read from the file
\item \code{scale_table} (data.table): if everything worked, the scale table
}
}
\description{
Creates a scale table from a given \code{.scl} file
}
\examples{

# a file with ratios specified in cents
alpha <- sclfile_scale_table(system.file(
  "test_scl_files/carlos_alpha.scl",
  package = "consonaR"
))
if (alpha$status == "Oll Korrect") {
  print(alpha$scale_table)
} else {
  print(alpha$status)
}

# a file with ratios specified as vulgar fractions
harm <- sclfile_scale_table(system.file(
  "test_scl_files/carlos_harm.scl",
  package = "consonaR"
))
if (harm$status == "Oll Korrect") {
  print(harm$scale_table)
} else {
  print(harm$status)
}

# another file with ratios specified as vulgar fractions
super <- sclfile_scale_table(system.file(
  "test_scl_files/carlos_super.scl",
  package = "consonaR"
))
if (super$status == "Oll Korrect") {
  print(super$scale_table)
} else {
  print(super$status)
}

# a file that doesn't exist
nosuch <- sclfile_scale_table(system.file(
  "carlo_scl_iles/carlos_harm.scl",
  package = "consonaR"
))
if (nosuch$status == "Oll Korrect") {
  print(nosuch$scale_table)
} else {
  print(nosuch$status)
}

}
