% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{sine_dissonance}
\alias{sine_dissonance}
\title{Sine Wave Dissonance}
\usage{
sine_dissonance(freq1, freq2, loud1, loud2)
}
\arguments{
\item{freq1}{frequency in Hertz of the first sine wave}

\item{freq2}{frequency in Hertz of the second sine wave}

\item{loud1}{loudness of the first sine wave}

\item{loud2}{loudness of the second sine wave}
}
\value{
the dissonance value
}
\description{
Calculates the dissonance function for two sine waves
}
\details{
This algorithm comes from Appendix E of
\insertCite{sethares2013tuning}{consonaR}
}
\references{
\insertAllCited{}
}
