% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{et_scale_table}
\alias{et_scale_table}
\title{Create Equal-Tempered Scale Table}
\usage{
et_scale_table(period = 2, divisions = 12, root_freq = 440/(2^(9/12)))
}
\arguments{
\item{period}{The period - default is 2, for an octave}

\item{divisions}{Number of degrees in the scale - default is 12}

\item{root_freq}{root frequency of the scale - default is middle C:
440 / (2 ^ (9 / 12))}
}
\value{
a \code{data.table} with six columns:
\itemize{
\item \code{degree}: scale degree from zero to (number of notes) - 1
\item \code{ratio_cents}: the ratio in cents (hundredths of a semitone)
\item \code{ratio_frac}: the ratio as a vulgar fraction (character). The ratios
for this type of scale are usually irrational, so this is an approximation,
computed by \code{fractional::fractional}.
\item \code{frequency}: frequency of the note given the \code{root_freq}
parameter
\item \code{interval_cents}: interval between this note and the previous note
in cents
\item \code{interval_frac}: interval between this note and the previous note
as a vulgar fraction
}
}
\description{
Creates a scale table for equal divisions of a specified
period.
}
\examples{

print(vanilla <- et_scale_table()) # default is 12EDO, of course

# 19-EDO
print(edo19 <- et_scale_table(2.0, 19))

# 31-EDO
print(edo31 <- et_scale_table(2.0, 31))

# equal-tempered Bohlen-Pierce
print(bohlen_pierce_et <- et_scale_table(3.0, 13))

# Carlos Alpha
print(carlos_alpha <- et_scale_table(1.5, 9))
}
